+++ server-rust/src/api/mod.rs
@@ -10,6 +10,7 @@ pub mod script;
 pub mod storage;
 pub mod stream;
 pub mod system_log;
+pub mod system_time;
 pub mod system_update;
 pub mod tailscale;
 pub mod vm;